[Client] Trying to fix the binary directory selection

+++ illacommon/src/main/java/illarion/common/util/DirectoryManager.java
@@ -19,6 +19,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -61,12 +62,19 @@ public final class DirectoryManager {
     private final Path workingDirectory;
 
     /**
+     * The binary directory that got selected.
+     */
+    @Nullable
+    private Path binaryDirectory;
+
+    /**
      * Private constructor to ensure that only the singleton instance exists.
      */
     @SuppressWarnings("nls")
     private DirectoryManager() {
         String installationDir = System.getProperty("org.illarion.install.dir");
         workingDirectory = Paths.get((installationDir == null) ? "." : installationDir);
+        binaryDirectory = null;
 
         Path userDir = getDirectory(Directory.User);
         if (Files.isRegularFile(userDir)) {
@@ -105,24 +113,34 @@ public final class DirectoryManager {
         throw new IllegalArgumentException("Parameter 'dir' was set to an illegal value: " + dir);
     }
 
+    @Nonnull
     private Path getBinaryDirectory() {
-        Path firstChoice = workingDirectory.resolve("bin");
-        if (!Files.exists(firstChoice)) {
-            try {
-                return Files.createDirectories(firstChoice);
-            } catch (IOException ignored) {
-                // not accessible
+        if (binaryDirectory == null) {
+            Path firstChoice = workingDirectory.resolve("bin");
+            if (!Files.exists(firstChoice)) {
+                try {
+                    return Files.createDirectories(firstChoice);
+                } catch (IOException ignored) {
+                    // not accessible
+                }
+            }
+            if (Files.isDirectory(firstChoice)) {
+                try {
+                    Path temporaryTestFile = firstChoice.resolve("writing.test");
+                    if (Files.exists(temporaryTestFile)) {
+                        Files.delete(temporaryTestFile);
+                    }
+                    Path newTempFile = Files.createFile(temporaryTestFile);
+                    Files.delete(newTempFile);
+                    binaryDirectory = firstChoice;
+                } catch (IOException ignored) {
+                }
+            }
+            if (binaryDirectory == null) {
+                binaryDirectory = getDirectory(Directory.User).resolve("bin");
             }
         }
-        if (Files.isDirectory(firstChoice)) {
-            try {
-                Path newTempFile = Files.createTempFile(firstChoice, "writing", ".text");
-                Files.delete(newTempFile);
-                return firstChoice;
-            } catch (IOException ignored) {
-            }
-        }
-        return getDirectory(Directory.User).resolve("bin");
+        return binaryDirectory;
     }
 
     @Nonnull
